requirements.txt
google-cloud-aiplatform
google-adk
google-cloud-bigquery
pydantic
python-dotenv
